Fix my brain damage for test object inheritence
